requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.4
sqlalchemy==2.0.23
python-dotenv==1.0.0

# Cloud Provider SDKs
boto3==1.28.62
google-cloud-billing==1.11.5
azure-mgmt-costmanagement==4.0.0
azure-identity==1.14.1

# Data Visualization & Dashboard
streamlit==1.27.2
plotly==5.17.0
matplotlib==3.7.3

# ML & Forecasting
prophet==1.1.4
scikit-learn==1.3.1

# Utilities
tqdm==4.66.1
pytest==7.4.2